solved space issue; vi ricorda still wip

+++ common.c
@@ -74,7 +74,7 @@ typedef struct track
     // target channel associated with this track
     channel *channel;
     // the sequence of commands associated with this track
-    uint16_t *sequence;
+    uint8_t *sequence;
     // the length of the command sequence
     uint16_t sLength;
     // the current position of the command sequence
